Visibility added, cleaned up a little.
Added loops to introduce the bricks on the screen. Using paint to loop
through for each rectangle drawn. (Easier for now to just use a
rectangle object for each brick as opposed to a sprite, we can add
sprites later once the physics works.) For now, running this gives a
flat, non moving image of the paddle, the ball, and some bricks.