Big Changes to overall layout and structure
Biggest change is the addition of the subclass "Ball". Current state of
the project: the ball moves (and bounces off borders of the screen, but
not yet the paddle or bricks), also using left or right to move will
move the paddle (although its somewhat rough and not as smooth as i
would like).